Render stacktraces properly Closes: #47921
